--- .vscode/settings.json
@@ -1,15 +1,0 @@
-{
-    "java.project.referencedLibraries": [
-        "lib/**/*.jar",
-        "JavaFX/javafx.base.jar",
-        "JavaFX/javafx.controls.jar",
-        "JavaFX/javafx.fxml.jar",
-        "JavaFX/javafx.graphics.jar",
-        "JavaFX/javafx.media.jar",
-        "JavaFX/javafx.swing.jar",
-        "JavaFX/javafx.web.jar",
-        "JavaFX/javafx-swt.jar"
-    ]
-
-    
-}
