+++ GraafinenLiittyma/src/main/java/com/example/PrimaryController.java
@@ -1,0 +1,12 @@
+package com.example;
+
+import java.io.IOException;
+import javafx.fxml.FXML;
+
+public class PrimaryController {
+
+    @FXML
+    private void switchToSecondary() throws IOException {
+        App.setRoot("secondary");
+    }
+}
