+++ GraafinenLiittyma/src/main/java/com/example/SecondaryController.java
@@ -1,0 +1,12 @@
+package com.example;
+
+import java.io.IOException;
+import javafx.fxml.FXML;
+
+public class SecondaryController {
+
+    @FXML
+    private void switchToPrimary() throws IOException {
+        App.setRoot("primary");
+    }
+}
