+++ GraafinenLiittyma/src/main/java/module-info.java
@@ -1,0 +1,7 @@
+module com.example {
+    requires javafx.controls;
+    requires javafx.fxml;
+
+    opens com.example to javafx.fxml;
+    exports com.example;
+}
