--- JavaFX/javafx.properties
@@ -1,3 +1,0 @@
-javafx.version=18.0.2
-javafx.runtime.version=18.0.2+2
-javafx.runtime.build=2
